requirements.txt
beautifulsoup4~=4.9.3
requests
requests_cache
pytz~=2020.4
pandas
pyyaml~=5.3.1

selenium~=3.141.0
setuptools~=50.3.2
